rename(components): 컴포넌트 구조 지역화
