+++ src/app/(domain)/community/_components/Post.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import PostsProps from '@/types/community/posts';
 import Link from 'next/link';
-import User from '@/app/_components/community/User';
+import User from '@/app/(domain)/community/_components/User';
 
 function Post({ ...props }: PostsProps) {
   return (
