+++ src/app/(domain)/community/_components/SearchBar.tsx
