+++ src/app/(domain)/community/_components/Tabs.tsx
