+++ src/app/(domain)/community/_components/comment/CommentBar.tsx
