+++ src/app/(domain)/community/page.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import Post from '@/app/_components/community/Post';
-import SearchBar from '@/app/_components/community/SearchBar';
-import Tabs from '@/app/_components/community/Tabs';
 import getPosts from '@/app/(domain)/community/api/getPosts';
 import { randomUUID } from 'crypto';
-import Floating from '@/app/_components/ui/buttons/Floating';
+import Floating from '@/app/ui/buttons/Floating';
+import Post from '@/app/(domain)/community/_components/Post';
+import SearchBar from '@/app/(domain)/community/_components/SearchBar';
+import Tabs from '@/app/(domain)/community/_components/Tabs';
 
 async function page() {
   const posts = await getPosts();
