+++ src/app/my-page/_components/Activity.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import Image from 'next/image';
-import Heading from '@/app/_components/my-page/Heading';
+import Heading from '@/app/my-page/_components/Heading';
 import { right } from '$/my-page';
 import { ACTIVITY_TITLE, ACTIVITY_LIST } from '@/constants/my-page/activity';
 import { randomUUID } from 'crypto';
