+++ src/app/my-page/_components/privacy/Content.tsx
