+++ src/app/my-page/layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from '@/app/_components/ui/common/Navbar';
+import Navbar from '@/app/ui/common/Navbar';
 
 function layout({ children }: { children: React.ReactNode }) {
   return (
