+++ src/app/signup/_components/Banner.tsx
