+++ src/app/signup/page.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import Banner from '@/app/_components/signup/Banner';
-import BirthInput from '@/app/_components/signup/BirthInput';
-import Gender from '@/app/_components/signup/Gender';
-import Button from '@/app/_components/ui/buttons/Button';
+import Button from '@/app/ui/buttons/Button';
 import { SIGNUP_BUTTON } from '@/constants/signup';
 import useSignupState from '@/store/signup/signupStore';
+import Banner from '@/app/signup/_components/Banner';
+import Gender from '@/app/signup/_components/Gender';
 
 function Page() {
   const { isValidBirth, isValidGender } = useSignupState();
@@ -15,7 +14,6 @@ function Page() {
     <section className="flex flex-col justify-between bg-white h-full mb-safe">
       <div>
         <Banner />
-        <BirthInput />
         <Gender />
       </div>
       <div className="mx-4 pb-4">
