+++ src/app/ui/common/Title.tsx
@@ -1,5 +1,5 @@
 import TEAM from '@/constants/team';
-import { rock } from '@/app/_components/ui/fonts';
+import { rock } from '@/app/ui/fonts';
 
 function Title() {
   return (
