+++ src/app/write/_components/PhotoBooth.tsx
